change size to size_t

+++ _examples/basic_c_example/main.c
@@ -3,7 +3,7 @@
 #include <unistd.h>
 #include <string.h>
 
-static void newMessage(char *instance, uint8_t *msg, int len, char *id)
+static void newMessage(char *instance, uint8_t *msg, size_t len, char *id)
 {
 	//this message can be deserialized here from something like json or whatever encoding you like
